Move PagedResult model to BillTrack.Core.Models.WebApi
