+++ BillTrack.Api/Features/Employees/Endpoints/GetAllEmployees.cs
@@ -1,7 +1,6 @@
-using BillTrack.Core.Contracts;
 using BillTrack.Core.Contracts.Employee;
 using BillTrack.Core.Interfaces.Services;
-using BillTrack.Core.Models;
+using BillTrack.Core.Models.WebApi;
 using BillTrack.Domain.Entities;
 using FastEndpoints;
 using IMapper = AutoMapper.IMapper;
