+++ BillTrack.Api/Features/Projects/Endpoints/GetAllProjects.cs
@@ -2,6 +2,7 @@ using BillTrack.Core.Contracts;
 using BillTrack.Core.Contracts.Project;
 using BillTrack.Core.Interfaces.Services;
 using BillTrack.Core.Models;
+using BillTrack.Core.Models.WebApi;
 using BillTrack.Domain.Entities;
 using FastEndpoints;
 using IMapper = AutoMapper.IMapper;
